requirements.txt
certifi==2017.4.17
chardet==3.0.3
click==6.7
DateTime==4.2
Flask==0.12.2
Flask-PyMongo==0.5.1
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
jsonify==0.5
MarkupSafe==1.0
pprint==0.1
pymongo==3.4.0
python-dotenv==0.6.4
pytz==2017.2
requests==2.17.3
urllib3==1.21.1
Werkzeug==0.12.2
zope.interface==4.4.1
